# calculate lt