--- a/src/likelihood.jl
+++ b/src/likelihood.jl
@@ -24,41 +24,36 @@ function calculate_LL!(myAmica::MultiModelAmica)
 end
 
 #Update loop for Lt and u (which is saved in z). Todo: Rename
-function loopiloop!(myAmica::SingleModelAmica)
-	(n,N) = size(myAmica.source_signals)
-	m = myAmica.m
-	Q = zeros(m,N)
-	@debug (:prop,myAmica.learnedParameters.proportions[1,1],:shape,myAmica.learnedParameters.shape[1],:y,myAmica.y[1])
+function loopiloop!(myAmica::SingleModelAmica, y_rho)
+	(n,_) = size(myAmica.source_signals)
+	
 	for i in 1:n
-		Q = calculate_Q(myAmica,Q,i) # myAmica.Q
+		Q = calculate_Q(myAmica,i, y_rho) # myAmica.Q
 		calculate_u!(myAmica,Q,i) # myAmica.z
-		@debug (:z,myAmica.z[1,1:5,1])
 		calculate_Lt!(myAmica,Q) # myAmica.Q
 	end
-	@debug (:Q,Q[1,1])
 end
 
-function loopiloop!(myAmica::MultiModelAmica)
+function loopiloop!(myAmica::MultiModelAmica, y_rho)
 	M = size(myAmica.models,1)
 	(n,_) = size(myAmica.models[1].source_signals)
 
 	for h in 1:M #run along models
-		
 		Threads.@threads for i in 1:n #run along components
-			Q = calculate_Q(myAmica.models[h], i)
+			Q = calculate_Q(myAmica.models[h], i, y_rho)
 			calculate_u!(myAmica.models[h], @view(Q[:, n, :]),i)
 			calculate_Lt!(myAmica.models[h], @view(Q[:, n, :]))
 		end
 	end
 end
 
-function calculate_Q(myAmica::SingleModelAmica, i)
+@views function calculate_Q(myAmica::SingleModelAmica, i, y_rho)
 	(n,N) = size(myAmica.source_signals)
 	m = myAmica.m
 	Q = zeros(m,N)
 	
 	for j in 1:m
-		Q[j,:] .= log(myAmica.learnedParameters.proportions[j,i]) + 0.5 * log(myAmica.learnedParameters.scale[j,i]) .+ logpfun(myAmica.y[i,:,j], myAmica.learnedParameters.shape[j,i])
+		Q[j,:] .= log(myAmica.learnedParameters.proportions[j,i]) + 0.5 * log(myAmica.learnedParameters.scale[j,i]) .+ logpfun(myAmica.learnedParameters.shape[j,i], y_rho[i, :, j])
 	end
 
 	return Q
@@ -75,7 +70,7 @@ end
 end
 
 #Applies location and scale parameter to source signals (per generalized Gaussian)
-function calculate_y!(myAmica::SingleModelAmica)
+@views function calculate_y!(myAmica::SingleModelAmica)
 	for j in 1:myAmica.m
 		myAmica.y[:,:,j] .= sqrt.(myAmica.learnedParameters.scale[j,:]) .* (myAmica.source_signals[:,:] .- myAmica.learnedParameters.location[j,:])
 	end
@@ -91,19 +86,6 @@ end
 		end
 	end
 end
-
-
-
-#Calculates densities for each generalized Gaussian j. Currently used my MultiModelAmica too
-function calculate_Q(myAmica::SingleModelAmica, Q, i)
-	m = size(myAmica.learnedParameters.scale, 1) #m = number of GGs, can't use myAmica.m in case this gets used my MultiModelAmica
-
-	for j in 1:m
-		Q[j,:] = log(myAmica.learnedParameters.proportions[j,i]) + 0.5*log(myAmica.learnedParameters.scale[j,i]) .+ logpfun(myAmica.y[i,:,j],myAmica.learnedParameters.shape[j,i])
-	end
-	return Q
-end
-
 
 #Calculates Likelihood for each time sample and for each ICA model
 function calculate_Lt!(myAmica::SingleModelAmica,Q)